Modifications after 2nd skype meeting (mainly derotation and derot_angle related)
