+++ vip/calib/__init__.py
@@ -6,6 +6,7 @@ from badframes import *
 from cosmetics import *
 from cosmetics_ifs import *
 from derotation import *
+from parangles import *
 from recentering import *
 from rescaling import *
 from subsampling import *
